+++ conftest.py
@@ -1,4 +1,6 @@
 import pytest
+import matplotlib
+matplotlib.use("Agg")
 
 
 def pytest_addoption(parser):
