+++ spyro/io/basicio.py
@@ -11,6 +11,12 @@ import segyio
 import glob
 import os
 import warnings
+from ..tools.version_control import is_firedrake_new
+
+
+if is_firedrake_new() is False:
+    from firedrake.__future__ import interpolate
+    fire.interpolate = interpolate
 
 
 def delete_tmp_files(wave):
