+++ spyro/meshing/meshing_operations.py
@@ -2,6 +2,12 @@ import firedrake as fire
 import numpy as np
 from spyro.utils.error_management import value_parameter_error
 from spyro.utils.eval_functions_to_ufl import generate_ufl_functions
+from ..tools.version_control import is_firedrake_new
+
+
+if is_firedrake_new() is False:
+    from firedrake.__future__ import interpolate
+    fire.interpolate = interpolate
 
 
 class MeshOps():
