+++ spyro/receivers/Receivers.py
@@ -2,6 +2,11 @@ from firedrake import *  # noqa: F403
 from spyro.receivers.dirac_delta_projector import Delta_projector
 from ..utils.typing import WaveType
 import numpy as np
+from ..tools.version_control import is_firedrake_new
+
+
+if is_firedrake_new() is False:
+    from firedrake.__future__ import interpolate
 
 
 class Receivers(Delta_projector):
