+++ spyro/tools/habc_tools.py
@@ -5,6 +5,11 @@ from numpy import clip, where
 from spyro.domains.space import create_function_space
 from spyro.utils.error_management import value_parameter_error
 from spyro.utils.eval_functions_to_ufl import generate_ufl_functions
+from ..tools.version_control import is_firedrake_new
+
+
+if is_firedrake_new() is False:
+    from firedrake.__future__ import interpolate
 
 
 def generate_conditional_value_for_layer(domain_dim, mesh, dimension,
