+++ spyro/utils/__init__.py
@@ -8,7 +8,6 @@ from .utils import (
     write_hdf5_velocity_model,
     get_real_shot_record,
     get_real_shot_step,
-    is_firedrake_new,
 )
 from .analytical_solution_nodal import nodal_homogeneous_analytical
 from .velocity_to_grid import velocity_to_grid, change_scalar_field_resolution
@@ -28,5 +27,4 @@ __all__ = [
     "write_hdf5_velocity_model",
     "get_real_shot_record",
     "get_real_shot_step",
-    "is_firedrake_new",
 ]
